♻️(recruteur-presentation) split views
